paging and kernel context switch working

+++ include/stdlib.h
@@ -6,7 +6,7 @@
 int main(int argc, char *argv[], char *envp[]);
 void exit(int status);
 
-void *malloc(size_t size);
+void *malloc(int size);
 void free(void *ptr);
 
 #endif
